melhorando delete do usuário + laçamento de exceções

+++ src/services/users/delete.ts
@@ -1,24 +1,25 @@
 import {BaseService} from '@services/base_service';
+import bcrypt from 'bcrypt';
+import { NotAuthrizedError, NotFoundError } from '@shared/errors';
 
 class UserDelete extends BaseService{
   constructor(){
     super()
   }
 
-  async call(id){
+  async call(password: string, id: number){
     let user = await this.prisma.user.findUnique({
       where: { id }
     });
 
-    console.log(user);
-    
-
     if(!user){
-      throw {
-        name: 'Record not found.',
-        message: `Usuário com id ${id} não foi encontrado`
-      }
+      throw new NotFoundError(`Usuário com id ${id} não foi encontrado`);
     }
+
+    if(!await bcrypt.compare(password, user.password)){      
+      throw new NotAuthrizedError(`Operação não autorizada`);
+    }
+    
 
     try {
       const user = await this.prisma.user.delete({
